changed foreign keys types to bigInteger

+++ melolib/database/migrations/2023_02_19_132737_create_bands_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('bands', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('cover_id');
+            $table->bigInteger('cover_id');
             $table->date('date_creation');
             $table->string('country');
             $table->timestamps();
