+++ melolib/database/migrations/2023_02_19_133704_create_band_artists_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('band_artists', function (Blueprint $table) {
             $table->id();
-            $table->integer('band_id');
-            $table->integer('artist_id');
+            $table->bigInteger('band_id');
+            $table->bigInteger('artist_id');
             $table->timestamps();
         });
     }
